% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat.ants-package.R
\docType{package}
\name{nat.ants-package}
\alias{nat.ants}
\alias{nat.ants-package}
\title{nat.ants: Bridge neuroanatomy toolbox ('nat') and 'ANTs' registration tools}
\description{
Integrate the neuroanatomy toolbox R package 'nat' and its
    many auxiliary packages with the Advanced Normalization Tools ('ANTs')
    registration toolkit developed at the University of Pennsylvania. This
    enables 'ANTs' registrations to be used by 'nat' to transform points or
    image data.
}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)

}
\keyword{internal}
