% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{xformpoints.antsreg}
\alias{xformpoints.antsreg}
\title{Map points using ANTs registrations}
\usage{
\method{xformpoints}{antsreg}(reg, points, ...)
}
\arguments{
\item{reg}{A \code{\link{antsreg}} object specifying one or more ANTs
registrations.}

\item{points}{An Nx3 matrix of 3D points to transform}

\item{...}{Additional arguments passed to \code{antsApplyTransformsToPoints}}
}
\value{
An Nx3 matrix of transformed points
}
\description{
Map points using ANTs registrations
}
\details{
See \code{\link{antsreg}} for details of the ordering of
  registrations.
}
\seealso{
Other antsreg: \code{\link{antsreg}}
}
\concept{antsreg}
